refactor(ws): split broker/ws.rs into ws/mod.rs + ws/rest.rs (#359)
Convert broker/ws.rs (1304 lines) into a ws/ submodule. WS-specific
code (upgrade handlers, session lifecycle, WsAppState, DaemonWsState,
create_router/create_daemon_router) stays in ws/mod.rs. All REST handler
functions and their request/query structs (api_join, api_send, api_poll,
api_query, api_health, daemon_api_*, build_query_filter,
apply_query_filter) move to ws/rest.rs.

create_router() and create_daemon_router() in mod.rs now reference
rest::api_* handlers by path. broker/mod.rs is unchanged — Rust resolves
ws/ submodule automatically.

While splitting, update 4 remaining direct status_map lock sites to use
the RoomState accessor methods added in #353:
- run_ws_session: set_status() + remove_status() instead of map.insert/remove
- api_health: status_count() instead of status_map.lock().await.len()
- daemon_api_health: status_count() on each room

No behavior changes. No wire format changes.
